+++ src/test/java/nextstep/subway/member/acceptance/MemberAcceptanceTest.java
@@ -1,0 +1,81 @@
+package nextstep.subway.member.acceptance;
+
+import static nextstep.subway.member.acceptance.MemberRestAssured.회원_삭제_요청;
+import static nextstep.subway.member.acceptance.MemberRestAssured.회원_생성을_요청;
+import static nextstep.subway.member.acceptance.MemberRestAssured.회원_정보_수정_요청;
+import static nextstep.subway.member.acceptance.MemberRestAssured.회원_정보_조회_요청;
+import static org.assertj.core.api.Assertions.assertThat;
+
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+import nextstep.subway.AcceptanceTest;
+import nextstep.subway.member.dto.MemberRequest;
+import nextstep.subway.member.dto.MemberResponse;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.http.HttpStatus;
+
+public class MemberAcceptanceTest extends AcceptanceTest {
+    private static final String EMAIL = "email@email.com";
+    private static final String PASSWORD = "password";
+    private static final String NEW_EMAIL = "newemail@email.com";
+    private static final String NEW_PASSWORD = "newpassword";
+    private static final int AGE = 20;
+    private static final int NEW_AGE = 21;
+    private MemberRequest 정상회원_등록_요청;
+
+    @BeforeEach
+    void setup() {
+        정상회원_등록_요청 = new MemberRequest(EMAIL, PASSWORD, AGE);
+    }
+
+    @DisplayName("회원 정보를 관리한다.")
+    @Test
+    void manageMember() {
+        // when
+        ExtractableResponse<Response> createResponse = 회원_생성을_요청(정상회원_등록_요청);
+        // then
+        회원_생성됨(createResponse);
+
+        // when
+        ExtractableResponse<Response> findResponse = 회원_정보_조회_요청(createResponse);
+        // then
+        회원_정보_조회됨(findResponse, EMAIL, AGE);
+
+        // when
+        ExtractableResponse<Response> updateResponse = 회원_정보_수정_요청(createResponse, NEW_EMAIL, NEW_PASSWORD, NEW_AGE);
+        // then
+        회원_정보_수정됨(updateResponse);
+
+        // when
+        ExtractableResponse<Response> deleteResponse = 회원_삭제_요청(createResponse);
+        // then
+        회원_삭제됨(deleteResponse);
+    }
+
+    @DisplayName("나의 정보를 관리한다.")
+    @Test
+    void manageMyInfo() {
+
+    }
+
+    private static void 회원_생성됨(ExtractableResponse<Response> response) {
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
+    }
+
+    private static void 회원_정보_조회됨(ExtractableResponse<Response> response, String email, int age) {
+        MemberResponse memberResponse = response.as(MemberResponse.class);
+        assertThat(memberResponse.getId()).isNotNull();
+        assertThat(memberResponse.getEmail()).isEqualTo(email);
+        assertThat(memberResponse.getAge()).isEqualTo(age);
+    }
+
+    private static void 회원_정보_수정됨(ExtractableResponse<Response> response) {
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
+    }
+
+    private static void 회원_삭제됨(ExtractableResponse<Response> response) {
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
+    }
+}
