+++ src/test/java/nextstep/subway/member/acceptance/MemberRestAssured.java
@@ -1,0 +1,60 @@
+package nextstep.subway.member.acceptance;
+
+import io.restassured.RestAssured;
+import io.restassured.parsing.Parser;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+import nextstep.subway.member.dto.MemberRequest;
+import org.springframework.http.MediaType;
+
+public class MemberRestAssured {
+
+    public static ExtractableResponse<Response> 회원_생성되어_있음(MemberRequest memberRequest) {
+        return 회원_생성을_요청(memberRequest);
+    }
+
+    public static ExtractableResponse<Response> 회원_생성을_요청(MemberRequest memberRequest) {
+        RestAssured.defaultParser = Parser.JSON;
+
+        return RestAssured
+                .given().log().all()
+                .body(memberRequest)
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .when().post("/members")
+                .then().log().all()
+                .extract();
+    }
+
+    public static ExtractableResponse<Response> 회원_정보_조회_요청(ExtractableResponse<Response> response) {
+        String uri = response.header("Location");
+
+        return RestAssured
+                .given().log().all()
+                .accept(MediaType.APPLICATION_JSON_VALUE)
+                .when().get(uri)
+                .then().log().all()
+                .extract();
+    }
+
+    public static ExtractableResponse<Response> 회원_정보_수정_요청(ExtractableResponse<Response> response, String email, String password, Integer age) {
+        String uri = response.header("Location");
+        MemberRequest memberRequest = new MemberRequest(email, password, age);
+
+        return RestAssured
+                .given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(memberRequest)
+                .when().put(uri)
+                .then().log().all()
+                .extract();
+    }
+
+    public static ExtractableResponse<Response> 회원_삭제_요청(ExtractableResponse<Response> response) {
+        String uri = response.header("Location");
+        return RestAssured
+                .given().log().all()
+                .when().delete(uri)
+                .then().log().all()
+                .extract();
+    }
+}
